legalHelp/requirements.txt
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pillow>=9.0.0
regex>=2022.0.0
weasyprint>=59.0